Updated money type to integer for "price" and "payment"

+++ console/migrations/m190508_145756_create_ingridient_table.php
@@ -24,7 +24,7 @@ class m190508_145756_create_ingridient_table extends Migration
             // порция
             //'dose' => $this->float()->notNull(),
             // цена ингридиента
-            'price' => $this->money()->notNull(),
+            'price' => $this->integer()->notNull(),
         ],$tableOptions);
     }
